# MetaSwapV3/docs/openapi.yaml
openapi: 3.1.0
info:
  title: MetaSwap V3 Core API
  version: 1.0.6
servers:
  - url: http://127.0.0.1:8080
paths:
...
  /tokens:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [issuerId, symbol, maxSupply, issuePriceUsd]
              properties:
                issuerId: { type: string }
                symbol: { type: string }
                name: { type: string }
                maxSupply: { type: number }
                issuePriceUsd: { type: number }
                chains:
                  type: array
                  items: { type: string }
                contracts:
                  type: object
      responses:
        "201":
          description: Token registered
...
  /orders:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, symbol, side, amount]
              properties:
                userId: { type: string }
                symbol: { type: string }
                quoteAsset: { type: string }
                side: { enum: [buy, sell] }
                amount: { type: number }
      responses:
        "201":
          description: Order result
...